backend/requirement.txt
asgiref==3.8.1
django-cors-headers==4.7.0
django-filter==25.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
mysqlclient==2.2.7
nepali-datetime==1.0.8.3
pillow==11.2.1
python-dotenv==1.1.0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2